fixup(heimdallm): surface GetPRByGithubID errors + move test helper
Addresses bot review findings on PR #258.

- runReview (daemon/cmd/heimdallm/main.go): log GetPRByGithubID errors
  other than sql.ErrNoRows instead of silently discarding. Behaviour
  stays fail-open (three downstream defenses still cap worst-case cost),
  but operators can now diagnose transient SQLite problems.
- Move InsertStaleInFlight from inflight.go to a _test.go file in the
  same package. Production code can no longer accidentally depend on it;
  external test files in package store_test still see it at test compile
  time.
- Strengthen the fail-open comment at the Claim site to enumerate the
  layered defenses (HEAD-SHA fail-closed, circuit breaker cap,
  PublishedAt dedup) so the reasoning is visible to the next reviewer.
  Add a cross-reference from the trigger-path Claim site.
